RF: Move config inheritance to `siblings`
