+++ datalad/distribution/add_sibling.py
@@ -19,7 +19,6 @@ from os.path import basename
 from datalad.utils import assure_list
 from datalad.support.param import Parameter
 from datalad.support.constraints import EnsureStr, EnsureNone
-from datalad.support.annexrepo import AnnexRepo
 from datalad.support.network import RI
 from ..interface.base import Interface
 from datalad.interface.common_opts import publish_depends
@@ -28,7 +27,7 @@ from datalad.interface.common_opts import annex_wanted_opt
 from datalad.interface.common_opts import annex_group_opt
 from datalad.interface.common_opts import annex_groupwanted_opt
 from datalad.interface.common_opts import inherit_opt
-from datalad.distribution.dataset import EnsureDataset, Dataset, \
+from datalad.distribution.dataset import EnsureDataset, \
     datasetmethod, require_dataset
 from datalad.support.exceptions import InsufficientArgumentsError
 
@@ -169,8 +168,6 @@ class AddSibling(Interface):
 
         # define config var name for potential publication dependencies
         depvar = 'remote.{}.datalad-publish-depends'.format(name)
-        # and default pushes
-        dfltvar = "remote.{}.push".format(name)
 
         # collect existing remotes:
         already_existing = list()
@@ -222,36 +219,6 @@ class AddSibling(Interface):
         if pushurl:
             ds.repo.set_remote_url(name, repo_props['pushurl'], push=True)
 
-        if inherit:
-            # Adjust variables which we should inherit
-            delayed_super = _DelayedSuper(ds.repo)
-            publish_depends = AddSibling._inherit_config_var(
-                delayed_super, depvar, publish_depends)
-            publish_by_default = AddSibling._inherit_config_var(
-                delayed_super, dfltvar, publish_by_default)
-            # Copy relevant annex settings for the sibling
-            # makes sense only if current AND super are annexes, so it is
-            # kinda a boomer, since then forbids having a super a pure git
-            if isinstance(ds.repo, AnnexRepo) \
-                and isinstance(delayed_super.repo, AnnexRepo):
-                if annex_wanted is None:
-                    annex_wanted = AddSibling._inherit_annex_var(
-                        delayed_super, name, 'wanted'
-                    )
-                if annex_group is None:
-                    # I think it might be worth inheritting group regardless what
-                    # value is
-                    #if annex_wanted in {'groupwanted', 'standard'}:
-                    annex_group = AddSibling._inherit_annex_var(
-                        delayed_super, name, 'group'
-                    )
-                if annex_wanted == 'groupwanted' and annex_groupwanted is None:
-                    # we better have a value for the expression for that group
-                    annex_groupwanted = AddSibling._inherit_annex_var(
-                        delayed_super, name, 'groupwanted'
-                    )
-
-
     @staticmethod
     def _inherit_annex_var(ds, remote, cfgvar):
         var = getattr(ds.repo, 'get_%s' % cfgvar)(remote)
@@ -276,36 +243,3 @@ def _urljoin(base, url):
     return base + url if (base.endswith('/') or url.startswith('/')) else base + '/' + url
 
 
-class _DelayedSuper(object):
-    """A helper to delay deduction on super dataset until needed
-
-    But if asked and not found -- blow up
-    """
-
-    def __init__(self, repo):
-        self._child_dataset = Dataset(repo.path)
-        self._super = None
-
-    def __str__(self):
-        return str(self.super)
-
-    @property
-    def super(self):
-        if self._super is None:
-            # here we must analyze current_ds's super, not the super_ds
-            self._super = self._child_dataset.get_superdataset()
-            if not self._super:
-                raise RuntimeError(
-                    "Cannot determine super dataset for %s, thus "
-                    "cannot inherit anything" % self._child_dataset
-                )
-        return self._super
-
-    # Lean proxies going through .super
-    @property
-    def config(self):
-        return self.super.config
-
-    @property
-    def repo(self):
-        return self.super.repo
